backend/requirements-full.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
email-validator==2.1.2
jinja2==3.1.2
clarifai==2.6.2
google-generativeai==0.3.2
cloudinary==1.36.0

